requirements.txt
aniso8601==9.0.1
attrs==23.1.0
awscli==1.29.78
blinker==1.6.3
boto3==1.28.78
botocore==1.31.78
click==8.1.7
colorama==0.4.4
docutils==0.16
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Principal==0.4.0
flask-restx==1.2.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
PyJWT==2.8.0
pyrsistent==0.20.0
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==2.3.7
